created useResponsive hook, voronoiCells and Circles components, functions to return voronoi and scales
