+++ src/App.css
@@ -1,4 +1,4 @@
-.App {
+.app {
   height: 100vh;
   width: 100vw;
   text-align: center;
