+++ src/charts/ScatterPlot/ScatterPlot.tsx
@@ -1,8 +1,21 @@
 /** App.js */
-import React, { useState, useEffect, useRef } from "react";
-import useEnvEffect from '../../hooks/useEnvEffect';
-import ScatterPlotBody from "./ScatterPlotBody";
-import { ScatterProps } from "../../../types";
+import React, {useMemo} from "react";
+import * as d3 from "d3"
+import { useResponsive } from '../../hooks/useResponsive';
+import Axis from "../../components/ContinuousAxis"
+import Circle from "../../components/Circle"
+import { d3Voronoi } from '../../functionality/voronoi';
+import { xScaleDef } from '../../functionality/xScale';
+import { yScaleDef } from '../../functionality/yScale';
+import VoronoiCell from "../../components/VoronoiCell";
+import { ScatterProps, Data, AccessorFunc, ColorScale } from "../../../types"
+import {
+  getXAxisCoordinates,
+  getYAxisCoordinates,
+  getMargins,
+  inferXDataType,
+} from "../../utils";
+
 
 export default function ScatterPlot({
   data,
@@ -18,46 +31,106 @@ export default function ScatterPlot({
   yGrid = false,
   xAxisLabel,
   yAxisLabel,
+  colorScheme = d3.schemeCategory10,
 }: ScatterProps<string | number>): JSX.Element {
-  const anchor = useRef(null as unknown as SVGSVGElement)
-  const [windowSize, setWindowSize] = useState<[number, number]>([0, 0])
-  const [cHeight, setCHeight] = useState<number>(0)
-  const [cWidth, setCWidth] = useState<number>(0)
+  
+const {anchor, cHeight, cWidth}  = useResponsive();
 
-  function updateSize() {
-    setWindowSize([window.innerWidth, window.innerHeight])
-  }
+const margin = useMemo(
+  () => getMargins(xAxis, yAxis, xAxisLabel, yAxisLabel),
+  [xAxis, yAxis, xAxisLabel, yAxisLabel]
+)
 
-  // Set up an event listener on mount
-  useEnvEffect(() => {
-    window.addEventListener("resize", updateSize)
-    updateSize()
-    return () => window.removeEventListener("resize", updateSize)
-  }, [])
+const { xAxisX, xAxisY } = useMemo(
+  () => getXAxisCoordinates(xAxis, cHeight, margin),
+  [cHeight, xAxis, margin]
+)
 
-  useEffect(() => {
-    const container = anchor.current.getBoundingClientRect()
-    setCHeight(container.height)
-    setCWidth(container.width)
-  }, [windowSize])
+const { yAxisX, yAxisY } = useMemo(
+  () => getYAxisCoordinates(yAxis, cWidth, margin),
+  [cWidth, yAxis, margin]
+)
+
+const translate = `translate(${margin.left}, ${margin.top})`
+
+if (!xDataType) {
+  xDataType = inferXDataType(data[0], xKey);
+}
+
+let keys: string[] = [],
+  groups: d3.InternMap<any, any[]>
+const groupAccessor = (d: any) => d[groupBy ?? ""]
+groups = d3.group(data, groupAccessor)
+keys = Array.from(groups).map((group) => group[0])
+
+
+const xAccessor: AccessorFunc = xDataType === 'number' ? (d) => d[xKey] : (d) => new Date(d[xKey]);
+const yAccessor: AccessorFunc = (d) => d[yKey];
+
+const xScale = xScaleDef(data, xDataType, xAccessor, margin, cWidth);
+const yScale = yScaleDef(data, margin, cHeight, yAccessor);
+
+const voronoi = d3Voronoi(data, xScale, yScale, xAccessor, yAccessor, cHeight, cWidth, margin)
+
+const colorScale: ColorScale = d3.scaleOrdinal(colorScheme)
+colorScale.domain(keys)
 
   return (
     <svg ref={anchor} width={width} height={height}>
-      <ScatterPlotBody
-        height={cHeight}
-        width={cWidth}
-        data={data}
-        xKey={xKey}
-        xDataType={xDataType}
-        yKey={yKey}
-        groupBy={groupBy}
-        xAxis={xAxis}
-        yAxis={yAxis}
-        xGrid={xGrid}
-        yGrid={yGrid}
-        xAxisLabel={xAxisLabel}
-        yAxisLabel={yAxisLabel}
-      />
+       <g className="spbody" transform={translate}>
+      {yAxis && (
+        <Axis
+          x={yAxisX}
+          y={yAxisY}
+          yGrid={yGrid}
+          height={cHeight}
+          width={cWidth}
+          margin={margin}
+          scale={yScale}
+          type={yAxis}
+          label={yAxisLabel}
+        />
+      )}
+      {xAxis && (
+        <Axis
+          x={xAxisX}
+          y={xAxisY}
+          xGrid={xGrid}
+          height={cHeight}
+          width={cWidth}
+          margin={margin}
+          scale={xScale}
+          type={xAxis}
+          label={xAxisLabel}
+        />
+      )}
+      {data.map((element: any, i: number) =>
+        !groupBy ? (
+          <Circle
+            key={i}
+            cx={xScale(xAccessor(element))}
+            cy={yScale(yAccessor(element))}
+            r={5}
+            color="steelblue"
+          />
+        ) : (
+          <Circle
+            key={i}
+            cx={xScale(xAccessor(element))}
+            cy={yScale(yAccessor(element))}
+            r={5}
+            color={colorScale(element[groupBy])}
+          />
+        )
+      )}
+      {voronoi && (
+        <g className="voronoi-wrapper">
+          {data.map((_elem: Data, i: number) => (
+            <VoronoiCell key={i} fill='none' stroke="#ff1493" opacity={0.5} d={voronoi.renderCell(i)}/>
+          ))}
+        </g>
+      )}
+    </g>
     </svg>
   )
 }
